requirements.txt
opencv-python>=4.8.1.78
numpy>=1.24.0
Pillow>=9.0.0
